Move yaml to serialize
