+++ src/ansible_navigator/actions/open_file.py
@@ -9,12 +9,12 @@ from typing import Callable
 from typing import Dict
 from typing import List
 
-from .._yaml import human_dump
 from ..app_public import AppPublic
 from ..configuration_subsystem import ApplicationConfiguration
 from ..ui_framework import Interaction
 from ..ui_framework import Menu
 from ..utils import remove_dbl_un
+from ..utils.serialize import human_dump
 from ..utils.serialize import json_dump
 from . import _actions as actions
