+++ src/ansible_navigator/actions/write_file.py
@@ -3,11 +3,11 @@ import logging
 import os
 import re
 
-from .._yaml import human_dump
 from ..app_public import AppPublic
 from ..configuration_subsystem import ApplicationConfiguration
 from ..ui_framework import Interaction
 from ..utils import remove_dbl_un
+from ..utils.serialize import human_dump
 from ..utils.serialize import json_dump
 from . import _actions as actions
