+++ src/ansible_navigator/configuration_subsystem/configurator.py
@@ -9,12 +9,12 @@ from typing import List
 from typing import Tuple
 from typing import Union
 
-from .._yaml import SafeLoader
-from .._yaml import yaml
 from ..utils import ExitMessage
 from ..utils import ExitPrefix
 from ..utils import LogMessage
 from ..utils import oxfordcomma
+from ..utils.serialize import SafeLoader
+from ..utils.serialize import yaml
 from .definitions import ApplicationConfiguration
 from .definitions import Constants as C
 from .parser import Parser
