+++ tests/unit/ui_framework/test_colorize.py
@@ -5,8 +5,8 @@ import os
 
 from unittest.mock import patch
 
-from ansible_navigator._yaml import human_dump
 from ansible_navigator.ui_framework.colorize import Colorize
+from ansible_navigator.utils.serialize import human_dump
 
 
 SHARE_DIR = os.path.abspath(
