+++ tests/unit/utils/test_serialize_yaml.py
@@ -5,7 +5,7 @@ import pytest
 
 from yaml import Dumper
 
-import ansible_navigator._yaml as yaml_import
+import ansible_navigator.utils.serialize as yaml_import
 
 
 def test_check_yaml_imports():
